requirements.txt
# ─── Phase 1: Auth Extractor ───
playwright>=1.49.0
playwright-stealth>=1.0.6

# ─── Phase 2: Mission Controller API ───
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-cloud-firestore>=2.19.0
google-cloud-tasks>=2.17.0
google-auth>=2.36.0

# ─── Phase 3: Sniper Worker ───
google-cloud-secret-manager>=2.21.0
